+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - Added `EuiTabbedContent` ([#737](https://github.com/elastic/eui/pull/737))
 
 **Bug fixes**
+
 - Fixed `EuiTableRowCell` from overwriting its child element's `className` [#709](https://github.com/elastic/eui/pull/709)
+- Allow `EuiContextMenuPanel`s to update when their `children` changes ([#710](https://github.com/elastic/eui/pull/710))
 
 ## [`0.0.44`](https://github.com/elastic/eui/tree/v0.0.44)